Python notes update
