--- 1_python_basics/4_variables/multiple_variables.py
@@ -1,20 +1,0 @@
-# Python allows you to assign values to multiple variables in one line.
-x, y, z = "Orange", 3, "Cherry"
-print(x)
-print(y)
-print(z)
-
-print("")
-
-x = y = z = "Orange"
-print(x)
-print(y)
-print(z)
-
-print("")
-
-fruits = ["apple", "banana", "cherry"]
-x, y, z = fruits
-print(x)
-print(y)
-print(z)
